Fixed flickering by pushing our backbuffer copy till the end.

+++ SoftwareRasterizer3D/include/Models/Vertex.h
@@ -6,10 +6,10 @@
 class Vertex {
 public:
 	Vertex() = default;
-	Vertex(Vec3<float> pos) : position(Vec3<float>(pos)) {}
-	Vertex(float x, float y, float z) : position(Vec3<float>(x, y, z)) {}
+	Vertex(Vec3<int> pos) : position(Vec3<int>(pos)) {}
+	Vertex(int x, int y, int z) : position(Vec3<int>(x, y, z)) {}
 
-	Vec3<float>& GetPosition();
+	Vec3<int>& GetPosition();
 	Vec4<int>& GetColour();
 	Vec3<float>& GetNormal();
 	Vec2<float>& GetUV();
@@ -18,7 +18,7 @@ public:
 	void SetColour(int r, int g, int b, int a = 255);
 
 private:
-	Vec3<float> position;
+	Vec3<int> position;
 	Vec4<int> colour;
 	Vec3<float> normal;
 	Vec2<float> UV;
